% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOntologyNetwork.R
\name{createOntologyNetwork}
\alias{createOntologyNetwork}
\title{Create ontology network}
\usage{
createOntologyNetwork(
  ontology.id,
  weighted = FALSE,
  method = "jaccard",
  organism = "hsa"
)
}
\arguments{
\item{ontology.id}{A character \code{\link{vector}} of ontology ID from supported ontology databases.}

\item{weighted}{Weight calculated by \code{method}. If FALSE (default), no weight is calculated.}

\item{method}{Calculated similarity between 2 ontology terms based on the \code{method}. Available methods are:
\describe{
\item{jaccard}{Jaccard Index}
\item{intersect}{Number of intersecting genes}
}}

\item{organism}{Organism identifier (required for GO annotations). Default is "hsa". Currently human (\code{hsa}), mouse (\code{mmu}) and rat (\code{rno}) are supported. Full list of organism IDs can be found [here](https://www.genome.jp/kegg/catalog/org_list.html).}
}
\value{
A \code{\link{data.frame}} with columns:
\describe{
\item{source}{Source node}
\item{target}{Target node}
\item{weight}{Weight of edge}
}
}
\description{
Creates a network of ontology terms based on the \code{method} provided.
}
\examples{
ontology.id <- sample_data$GOBP$ID[1:10]
createOntologyNetwork(ontology.id, method = "jaccard")
}
