% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordCloud.R
\name{createWordCloud}
\alias{createWordCloud}
\title{Create Word Cloud}
\usage{
createWordCloud(
  oc,
  by = c("ontology_cluster", "link_community")[1],
  excludeWord = NULL
)
}
\arguments{
\item{oc}{\code{ontoClust} object. The output of \code{\link{getOntoClust}}}

\item{by}{Create the word cloud from either "ontology cluster" or "link community". Default is "ontology_cluster"}

\item{excludeWord}{Words to exclude from word-cloud (e.g. regulation)}
}
\value{
A \code{\link{data.frame}} with columns "word", "freq" and "group".
}
\description{
Create Word Cloud
}
\examples{
ontology.id <- sample_data$GOBP$ID[1:50]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
createWordCloud(oc)
}
