% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOntoDB.R
\name{detectOntoDB}
\alias{detectOntoDB}
\title{Detect ontology database from ontology ID}
\usage{
detectOntoDB(ontology.id, NA.if.no.match = TRUE)
}
\arguments{
\item{ontology.id}{A character vector with ontology IDs}

\item{NA.if.no.match}{If an ID is not present across all databases, corresponding element of vector will be assigned as NA (default: TRUE). If FALSE, function will throw an error and stop.}
}
\value{
A character vector with the name of the databases.
}
\description{
Checks each ID across "DO", "GO", "KEGG", "MKEGG" and "REACTOME" databases and returns the database name.
}
\examples{
ontology.ids <- lapply(sample_data, function(x){head(x$ID, 3)})
ontology.ids <- do.call(c, ontology.ids)
ontology.ids
detectOntoDB(ontology.ids)

ontology.ids2 <- c(ontology.ids[1:3], "random-id-here")
ontology.ids2
detectOntoDB(ontology.ids2)
}
