% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordCloud.R
\name{.createWordCloud}
\alias{.createWordCloud}
\title{Engine for Create Word Cloud}
\usage{
.createWordCloud(ontologyDescriptions, excludeWord = NULL)
}
\arguments{
\item{ontologyDescriptions}{Description of ontology terms}

\item{excludeWord}{Words to exclude from word cloud}
}
\value{
A \code{\link{data.frame}} with columns "word", "freq"
}
\description{
Engine for Create Word Cloud
}
\examples{
\dontrun{
ontology.id <- sample_data$GOBP$ID[1:50]
x <- onto2description(ontology.id)
y <- x$description
.createWordCloud(y)
}
}
