% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOntoDB.R
\name{.detectOntoDB}
\alias{.detectOntoDB}
\title{Detect ontology database from ontology ID.}
\usage{
.detectOntoDB(ontology.id, NA.if.no.match = FALSE)
}
\arguments{
\item{ontology.id}{A character vector with ontology IDs}

\item{NA.if.no.match}{If an ID is not present across all databases, corresponding element of vector will be assigned as NA (default: TRUE). If FALSE, function will throw an error and stop.}
}
\value{
A character vector with the name of the databases.
}
\description{
Checks ontology ID across "DO", "GO", "KEGG", "MKEGG" and "REACTOME" databases and returns the database name. This is the engine function for \code{\link{detectOntoDB}} and checks one entry at a time.
}
