% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLC.R
\name{getLC}
\alias{getLC}
\title{Get Link Communities}
\usage{
getLC(network, hcmethod = "single")
}
\arguments{
\item{network}{A \code{\link{data.frame}} representing network with columns "source", "target" and "weight" (optional).}

\item{hcmethod}{Hierarchical clustering method. See \code{\link{hclust}} for details. Recommended to use "single" for unwighted networks and "ward.D2" for weighted networks.}
}
\value{
An object of class \code{\link{linkcomm}}. See \code{\link{getLinkCommunities}} for details.
}
\description{
Get Link Communities
}
\examples{
\dontrun{
ontology.id <- sample_data$GOBP$ID[1:100]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
lc
}
}
