% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntoClust.R
\name{getOntoClust}
\alias{getOntoClust}
\title{Get clusters of link communitites}
\usage{
getOntoClust(lc, minClusterSize = 3, verbose = TRUE)
}
\arguments{
\item{lc}{Output of \code{\link{getLC}}}

\item{minClusterSize}{Minimum number of members in a cluster. Default is 3.}

\item{verbose}{Print detailed output. Default is TRUE.}
}
\value{
A \code{\link{list}} object with
\item{\code{mat}}{Binary matrix. Rows: ontologies, columns: link communities}
\item{\code{ontoClust}}{data.frame with columns of ontology IDs and link communities}
\item{\code{dist}}{Distance matrix of \code{mat}}
\item{\code{hc}}{\code{\link{hclust}} object}
}
\description{
Get clusters of link communitites
}
\examples{
\dontrun{
ontology.id <- sample_data$GOBP$ID[1:100]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
}
}
