% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go2description.R
\name{go2description}
\alias{go2description}
\title{Gene Ontology (GO) to Description}
\usage{
go2description(go.id)
}
\arguments{
\item{go.id}{A vector of DO IDs}
}
\value{
A \code{\link{data.frame}} with GO IDs and their descriptions.
}
\description{
Gene Ontology (GO) to Description
}
\examples{
go.ids <- sample_data$GOBP$ID[1:3]
go2description(go.ids)

}
