% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go2gene.R
\name{go2gene}
\alias{go2gene}
\title{Gene Ontology (GO) to Gene ID}
\usage{
go2gene(go.id, organism = "hsa")
}
\arguments{
\item{go.id}{GO ID (e.g. GO:0045333)}

\item{organism}{Organism identifier. Currently human (\code{hsa}), mouse (\code{mmu}) and rat (\code{rno}) are supported. Full list of organism IDs can be found [here](https://www.genome.jp/kegg/catalog/org_list.html).}
}
\value{
A list object containing Gene IDs for every GO ID
}
\description{
Gene Ontology (GO) to Gene ID
}
\examples{
go.ids <- sample_data$GOBP$ID[1:3]
go2gene(go.ids)
}
