% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardIndex.R
\name{jaccardIndex}
\alias{jaccardIndex}
\title{Calculate Jaccard Index}
\usage{
jaccardIndex(x, y)
}
\arguments{
\item{x}{Set 1 \code{\link{vector}}}

\item{y}{Set 1 \code{\link{vector}}}
}
\value{
Jaccard Index
}
\description{
Calculates Jaccard Index (length of intersection divided by length of union) for 2 given sets.
}
\examples{
id.x <- sample_data$GOBP$ID[1]
id.y <- sample_data$GOBP$ID[2]

x <- go2gene(id.x)
y <- go2gene(id.y)

jaccardIndex(x, y)
}
