% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg2description.R
\name{kegg2description}
\alias{kegg2description}
\title{KEGG ID to Description}
\usage{
kegg2description(kegg.id)
}
\arguments{
\item{kegg.id}{A vector of KEGG IDs}
}
\value{
A \code{\link{data.frame}} with KEGG IDs and their descriptions.
}
\description{
KEGG ID to Description
}
\examples{
\dontrun{
kegg.ids <- sample_data$KEGG$ID[1:3]
kegg2description(kegg.ids)
}
}
