% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg2gene.R
\name{kegg2gene}
\alias{kegg2gene}
\title{KEGG to Gene ID}
\usage{
kegg2gene(kegg.id)
}
\arguments{
\item{kegg.id}{A vector of KEGG IDs}
}
\value{
A list object containing Gene IDs for every KEGG IDs
}
\description{
KEGG to Gene ID
}
\examples{
\dontrun{
kegg.ids <- sample_data$KEGG$ID[1:3]
kegg2gene(kegg.ids)
}

}
