% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkegg2description.R
\name{mkegg2description}
\alias{mkegg2description}
\title{MKEGG ID to Description}
\usage{
mkegg2description(mkegg.id)
}
\arguments{
\item{mkegg.id}{A vector of MKEGG IDs}
}
\value{
A \code{\link{data.frame}} with MKEGG IDs and its descriptions.
}
\description{
MKEGG ID to Description
}
\examples{
\dontrun{
mkegg.ids <- sample_data$MKEGG$ID[1:3]
mkegg2description(mkegg.ids)
}
}
