% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onto2description.R
\name{onto2description}
\alias{onto2description}
\title{Ontology ID to Ontology Descriptions}
\usage{
onto2description(id)
}
\arguments{
\item{id}{ID of any ontology term}
}
\value{
A \code{\link{data.frame}} with ontology IDs and their descriptions.
}
\description{
Ontology ID to Ontology Descriptions
}
\examples{
\dontrun{
id <- lapply(sample_data, function(x){head(x$ID)})
id <- do.call(c, id)
onto2description(id)
}
}
