% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onto2gene.R
\name{onto2gene}
\alias{onto2gene}
\title{Title}
\usage{
onto2gene(id)
}
\arguments{
\item{id}{ID of any ontology term}
}
\value{
A \code{\link{list}} object containing gene IDs for each ontology ID
}
\description{
Title
}
\examples{
\dontrun{
ids <- sample_data$KEGG$ID[1:5]
onto2gene(ids)
}
}
