% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOntoNetwork.R
\name{plotOntoNetwork}
\alias{plotOntoNetwork}
\title{Plot Network of Ontology Clusters}
\usage{
plotOntoNetwork(
  oc,
  layout = "kk",
  size_ontology = 3,
  size_label_ontology = 2,
  alpha_edge = 0.2,
  alpha_node = 1,
  facet_by = NULL
)
}
\arguments{
\item{oc}{\code{ontoClust} object. The output of \code{\link{getOntoClust}}}

\item{layout}{Network layout. See \code{\link{ggraph}} for details. Default is "kk". Some options are 'kk', 'dh', 'circle', 'gem', 'graphopt', 'grid', 'mds', 'randomly', 'fr', 'drl', 'lgl', 'nicely'}

\item{size_ontology}{Size of ontology (node) terms. Default is 3.}

\item{size_label_ontology}{Size of ontology (node) term labels. Default is 2.}

\item{alpha_edge}{Transparency parameter for edges (link communities). Default is 0.2.}

\item{alpha_node}{Transparency parameter for nodes (ontology clusters). Default is 1 (fully opaque).}

\item{facet_by}{Facets to draw. Default is NULL, thus no facets are drawn. It can be 'ontology_cluster' or 'link_community'}
}
\value{
\code{\link{ggraph}} plot
}
\description{
Plot Network of Ontology Clusters
}
\examples{
\dontrun{
ontology.id <- sample_data$GOBP$ID[1:100]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
plotOntoNetwork(oc)
}
}
