% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWordCloud.R
\name{plotWordCloud}
\alias{plotWordCloud}
\title{Plot Word Cloud}
\usage{
plotWordCloud(wc, n = Inf, freq = TRUE, norm.freq = TRUE)
}
\arguments{
\item{wc}{\code{\link{data.frame}} for word cloud with frequencies and group. The columns are
\describe{
\item{word}{Word (character)}
\item{freq}{Frequency of the word (numeric)}
\item{norm.freq}{Normalized frequency of the word (numeric). Normalization is done by division by the sum of frequencies}
\item{group}{ID of Ontology Cluster or Link Community (numeric)}
}}

\item{n}{Top "n" words to plot (numeric). Default is Inf (infinite), which uses all available words.}

\item{freq}{Use frequency of the words as size. Default TRUE.}

\item{norm.freq}{Use normalized frequency of the words as size. Default TRUE.}
}
\value{
A \code{\link{ggwordcloud}} object
}
\description{
Plot Word Cloud
}
\examples{
ontology.id <- sample_data$GOBP$ID[1:50]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
wc <- createWordCloud(oc)
plotWordCloud(wc, n = 10, freq = TRUE, norm.freq = TRUE)
}
