% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome2description.R
\name{reactome2description}
\alias{reactome2description}
\title{Reactome ID to Description}
\usage{
reactome2description(reactome.id)
}
\arguments{
\item{reactome.id}{A vector of Reactome IDs}
}
\value{
A \code{\link{data.frame}} with Reactome IDs and their descriptions.
}
\description{
Reactome ID to Description
}
\examples{
reactome.ids <- sample_data$REACTOME$ID[1:3]
reactome2description(reactome.ids)
}
