% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome2gene.R
\name{reactome2gene}
\alias{reactome2gene}
\title{Reactome to Gene ID}
\usage{
reactome2gene(reactome.id)
}
\arguments{
\item{reactome.id}{REACTOME ID (e.g. R-HSA-611105)}
}
\value{
A list object containing Gene IDs for every GO ID
}
\description{
Reactome to Gene ID
}
\examples{
reactome.ids <- sample_data$REACTOME$ID[1:3]
reactome2gene(reactome.ids)
}
