% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do2gene.R
\name{do2gene}
\alias{do2gene}
\title{Disease Ontology (DO) to Gene ID (ENTREZ)}
\usage{
do2gene(do.id)
}
\arguments{
\item{do.id}{A vector of DO IDs}
}
\value{
A list object containing Gene IDs for every DO ID
}
\description{
Disease Ontology (DO) to Gene ID (ENTREZ)
}
\examples{
do.ids <- sample_data$DO$ID[1:3]
do2gene(do.ids)
}
\seealso{
\code{\link{onto2gene}} for all ontology to gene conversions as well as \code{\link{go2gene}}, \code{\link{kegg2gene}}, \code{\link{reactome2gene}}.
}
