% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntoClust.R
\name{getOntoClust}
\alias{getOntoClust}
\title{Get clusters of link communitites}
\usage{
getOntoClust(lc, minClusterSize = 3)
}
\arguments{
\item{lc}{Output of \code{\link{getLC}}}

\item{minClusterSize}{Minimum number of members in a cluster. Default is 3.}
}
\value{

}
\description{
Get clusters of link communitites
}
\examples{
\dontrun{
ontology.id <- sample_data$GOBP$ID[1:100]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
}
}
