% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapOntoClust.R
\name{heatmapOntoClust}
\alias{heatmapOntoClust}
\title{Create a heatmap of ontology clusters}
\usage{
heatmapOntoClust(
  oc,
  clusterColors = NULL,
  clusterLC = FALSE,
  rowNames = c("id", "description")[2],
  filename = NA,
  silent = FALSE,
  ...
)
}
\arguments{
\item{oc}{Output from \code{\link{getOntoClust}}}

\item{clusterColors}{A character \code{\link{vector}} with length of clusters (rows). If NULL (default), \code{\link{rainbow}} colors will be used.}

\item{clusterLC}{Logical to cluster link communities (columns).}

\item{rowNames}{Display text for ontologies. It can be either
\describe{
\item{id}{Ontology ID}
\item{description}{Description of the ontology ID (default)}
}}

\item{filename}{Default NULL. If provided with .pdf or .png extensions, function generates a PDF or PNG output file.}

\item{silent}{Default FALSE. If TRUE, the function does not plot the heatmap to the graphics device.}

\item{...}{Additional parameters for \code{\link{pheatmap}}}
}
\value{
Pheatmap
}
\description{
Create a heatmap of ontology clusters
}
\examples{
ontology.id <- sample_data$GOBP$ID[1:50]
network <- createOntologyNetwork(ontology.id, method = "jaccard", weighted = FALSE)
lc <- getLC(network)
oc <- getOntoClust(lc)
heatmapOntoClust(oc)
}
